+++ server/impl/src/main/java/com/walmartlabs/concord/server/events/AbstractEventResource.java
@@ -21,6 +21,7 @@ package com.walmartlabs.concord.server.events;
  */
 
 import com.walmartlabs.concord.sdk.Constants;
+import com.walmartlabs.concord.server.ConcordApplicationException;
 import com.walmartlabs.concord.server.cfg.TriggersConfiguration;
 import com.walmartlabs.concord.server.org.project.ProjectDao;
 import com.walmartlabs.concord.server.org.triggers.TriggerEntry;
@@ -29,9 +30,15 @@ import com.walmartlabs.concord.server.process.Payload;
 import com.walmartlabs.concord.server.process.PayloadBuilder;
 import com.walmartlabs.concord.server.process.ProcessManager;
 import com.walmartlabs.concord.server.security.UserPrincipal;
+import com.walmartlabs.concord.server.security.ldap.LdapManager;
+import com.walmartlabs.concord.server.security.ldap.LdapPrincipal;
+import com.walmartlabs.concord.server.user.UserEntry;
+import com.walmartlabs.concord.server.user.UserManager;
+import com.walmartlabs.concord.server.user.UserType;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import javax.naming.NamingException;
 import java.io.IOException;
 import java.util.HashMap;
 import java.util.List;
@@ -49,16 +56,25 @@ public abstract class AbstractEventResource {
     private final ProjectDao projectDao;
     private final TriggerDefinitionEnricher triggerDefinitionEnricher;
     private final TriggersConfiguration triggersCfg;
+    private final UserManager userManager;
+    private final LdapManager ldapManager;
 
     public AbstractEventResource(ProcessManager processManager,
-                                 TriggersDao triggersDao, ProjectDao projectDao, TriggersConfiguration triggersCfg) {
-        this(processManager, triggersDao, projectDao, AS_IS_ENRICHER, triggersCfg);
+                                 TriggersDao triggersDao,
+                                 ProjectDao projectDao,
+                                 TriggersConfiguration triggersCfg,
+                                 UserManager userManager,
+                                 LdapManager ldapManager) {
+
+        this(processManager, triggersDao, projectDao, AS_IS_ENRICHER, triggersCfg, userManager, ldapManager);
     }
 
     public AbstractEventResource(ProcessManager processManager,
                                  TriggersDao triggersDao, ProjectDao projectDao,
                                  TriggerDefinitionEnricher enricher,
-                                 TriggersConfiguration triggersCfg) {
+                                 TriggersConfiguration triggersCfg,
+                                 UserManager userManager,
+                                 LdapManager ldapManager) {
 
         this.processManager = processManager;
         this.triggersDao = triggersDao;
@@ -66,6 +82,8 @@ public abstract class AbstractEventResource {
         this.triggerDefinitionEnricher = enricher;
         this.log = LoggerFactory.getLogger(this.getClass());
         this.triggersCfg = triggersCfg;
+        this.userManager = userManager;
+        this.ldapManager = ldapManager;
     }
 
     protected int process(String eventId, String eventName, Map<String, Object> conditions, Map<String, Object> event) {
@@ -87,8 +105,9 @@ public abstract class AbstractEventResource {
             args.put("event", event);
 
             try {
+                UserEntry initiator = getInitiator(t, event);
                 UUID orgId = projectDao.getOrgId(t.getProjectId());
-                UUID instanceId = startProcess(orgId, t.getProjectId(), t.getRepositoryId(), t.getEntryPoint(), t.getActiveProfiles(), args);
+                UUID instanceId = startProcess(orgId, t.getProjectId(), t.getRepositoryId(), t.getEntryPoint(), t.getActiveProfiles(), args, initiator);
                 log.info("process ['{}'] -> new process ('{}') triggered by {}", eventId, instanceId, t);
             } catch (Exception e) {
                 log.error("process ['{}', '{}', '{}'] -> error", event, eventName, t.getId(), e);
@@ -102,6 +121,31 @@ public abstract class AbstractEventResource {
         return triggersCfg.isDisableAll() || triggersCfg.getDisabled().contains(eventName);
     }
 
+    private UserEntry getInitiator(TriggerEntry trigger, Map<String, Object> event) {
+        boolean isUseInitiator = trigger.isUseInitiator();
+        if (!isUseInitiator) {
+            UserPrincipal initiator = UserPrincipal.assertCurrent();
+            return initiator.getUser();
+        }
+
+        String author = event.getOrDefault("author", "").toString();
+        return getOrCreateUser(author);
+
+    }
+
+    private UserEntry getOrCreateUser(String author) {
+        try {
+            LdapPrincipal p = ldapManager.getPrincipal(author);
+            if (p == null) {
+                throw new ConcordApplicationException("User not found: " + author);
+            }
+
+            return userManager.getOrCreate(author, UserType.LDAP);
+        } catch (NamingException e) {
+            throw new ConcordApplicationException("Error while retrieving LDAP data: " + e.getMessage(), e);
+        }
+    }
+
     private boolean filter(Map<String, Object> conditions, TriggerEntry t) {
         try {
             return EventMatcher.matches(conditions, t.getConditions());
@@ -111,21 +155,27 @@ public abstract class AbstractEventResource {
         }
     }
 
-    private UUID startProcess(UUID orgId, UUID projectId, UUID repoId, String entryPoint, List<String> activeProfiles, Map<String, Object> args) throws IOException {
-        UUID instanceId = UUID.randomUUID();
+    private UUID startProcess(UUID orgId,
+                              UUID projectId,
+                              UUID repoId,
+                              String entryPoint,
+                              List<String> activeProfiles,
+                              Map<String, Object> args,
+                              UserEntry initiator) throws IOException {
 
-        UserPrincipal initiator = UserPrincipal.assertCurrent();
+        UUID instanceId = UUID.randomUUID();
 
         Map<String, Object> request = new HashMap<>();
         if (activeProfiles != null) {
             request.put(Constants.Request.ACTIVE_PROFILES_KEY, activeProfiles);
         }
+
         if (args != null) {
             request.put(Constants.Request.ARGUMENTS_KEY, args);
         }
 
         Payload payload = new PayloadBuilder(instanceId)
-                .initiator(initiator.getId(), initiator.getUsername())
+                .initiator(initiator.getId(), initiator.getName())
                 .organization(orgId)
                 .project(projectId)
                 .repository(repoId)